Rename outfileType 'apollo-js' to 'js' for the JS client
The generated JS code accounts for 'relay', 'apollo', and plain JS. This
rename is to not make it seem like this only works for 'apollo'.

We can always add more specific generators in the future.
